+++ packages/crypto/packages/target-bundler/index.d.ts
@@ -1,2 +1,0 @@
-import { IrohaCryptoInterface } from '@iroha2/crypto-core'
-export declare const crypto: IrohaCryptoInterface