+++ packages/crypto/packages/target-bundler/index.js
@@ -1,5 +1,0 @@
-import * as wasm from "./wasm_pack_output";
-const crypto = wasm;
-export {
-  crypto
-};