+++ packages/crypto/packages/target-bundler/wasm_pack_output.js
@@ -1,3 +1,0 @@
-import * as wasm from "./wasm_pack_output_bg.wasm";
-export * from "./wasm_pack_output_bg.js";
-wasm.__wbindgen_start();