+++ packages/crypto/packages/target-node/index.ts
@@ -1,3 +1,0 @@
-import { IrohaCryptoInterface } from '@iroha2/crypto-core'
-import * as wasm from './wasm_pack_output'
-export const crypto: IrohaCryptoInterface = wasm